requirements.txt
pylint==1.7.4
python-telegram-bot==8.1.1
PyGreSQL==5.0.4
